scripts/requirements-dev.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
pydantic>=1.8.0
loguru>=0.5.3
joblib>=1.0.0
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5
kafka-python>=2.0.2
pyarrow>=6.0.0
fastparquet>=0.7.2
python-dotenv>=0.19.0
requests>=2.26.0
prometheus-client>=0.12.0
tabulate>=0.9.0
